requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
future==0.17.1
idna==2.8
importlib-metadata==0.17
isort==4.3.20
lazy-object-proxy==1.4.1
lxml==4.3.4
mccabe==0.6.1
more-itertools==7.0.0
numpy==1.16.4
packaging==19.0
pandas==0.24.2
pkg-resources==0.0.0
pluggy==0.12.0
py==1.8.0
pycryptodomex==3.8.1
pyjwkest==1.4.0
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.2
python-dateutil==2.8.0
python-docx==0.8.10
python-taiga==0.9.0
pytz==2019.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.1
zipp==0.5.1
